corrections square

+++ src/fr/fms/graphic/Square.java
@@ -1,23 +1,23 @@
 package fr.fms.graphic;
 
 public class Square extends Shape {
-	private int side;
+	private double side;
 
-	public Square(int side, int x, int y) {
+	public Square(double side, int x, int y) {
 		super(x, y);
 		this.setSide(side);
 	}
 
-	public Square(int side, Point center) {
+	public Square(double side, Point center) {
 		super(center);
 		this.setSide(side);
 	}
 
-	public int getSide() {
+	public double getSide() {
 		return side;
 	}
 
-	public void setSide(int side) {
+	public void setSide(double side) {
 		if (side < 0)
 			side = 1;
 		else
